Moved torch compile code into comfy_api so it can be used by custom nodes with a degree of confidence
