+++ comfy_api/torch_helpers/__init__.py
@@ -1,0 +1,5 @@
+from .torch_compile import set_torch_compile_wrapper
+
+__all__ = [
+    "set_torch_compile_wrapper",
+]
